Refactor Notification Email into activity
Refactored notification email editor so that Assignments and Quizzing will both use the same component. When the input field is changed, it will fire an event, which will the host component (d2l-activity-assignment-editor-submission-and-completion or d2l-activity-quiz-attempts-and-completion-editor) will hear and update the correct store (i.e. assignments or quizzing) with the new notification email value.
This component also has the input validation, so by switching quizzing to use this, it is going to automatically pick up the error validation.
